+++ tests/integration/integration_base.py
@@ -54,7 +54,14 @@ class IntegrationTestBase(unittest.TestCase):
                       'service_descriptions': [{
                           'url': 'https://raw.githubusercontent.com/OpenAPITools/openapi-generator/master/modules/openapi-generator-gradle-plugin/samples/local-spec/petstore-v3.0.yaml',
                           'rest_service_code': 'Petstore',
-                          'type': 'OPENAPI3'
+                          'type': 'OPENAPI3',
+                          'access': ['BUS'],
+                          'services': [
+                              {
+                                  'service_code': 'Petstore',
+                                  'access': ['BUS']
+                              }
+                          ]
                       }
                       ]
                   }]