+++ tests/integration/test_xrdsst.py
@@ -2,7 +2,7 @@ import urllib3
 
 from tests.integration.integration_base import IntegrationTestBase
 from tests.integration.integration_ops import IntegrationOpBase
-from tests.util.test_util import get_client, auth_cert_registration_global_configuration_update_received, waitfor
+from tests.util.test_util import get_client, auth_cert_registration_global_configuration_update_received, waitfor, get_service_clients
 from tests.util.test_util import get_service_description, assert_server_statuses_transitioned
 from xrdsst.controllers.base import BaseController
 from xrdsst.controllers.cert import CertController
@@ -123,19 +123,30 @@ class TestXRDSST(IntegrationTestBase, IntegrationOpBase):
             service_description = get_service_description(self.config, client_id)
             assert service_description["disabled"] is False
 
-    def step_add_service_access_rights(self):
+    def step_add_service_access(self, client_id):
         with XRDSSTTest() as app:
             service_controller = ServiceController()
             service_controller.app = app
             service_controller.load_config = (lambda: self.config)
-            service_controller.add_rights
+            service_controller.add_access
+            description = get_service_description(self.config, client_id)
+            service_clients = get_service_clients(self.config, description["services"][0]["id"])
+            assert len(service_clients) == 1
 
-    def step_update_service_parameters(self):
+    def step_update_service_parameters(self, client_id):
         with XRDSSTTest() as app:
             service_controller = ServiceController()
+            description = get_service_description(self.config, client_id)
+            assert description["services"][0]["timeout"] == 60
+            assert description["services"][0]["ssl_auth"] is False
+            assert description["services"][0]["url"] == 'http://petstore'
             service_controller.app = app
             service_controller.load_config = (lambda: self.config)
             service_controller.update_parameters
+            description = get_service_description(self.config, client_id)
+            assert description["services"][0]["timeout"] == 120
+            assert description["services"][0]["ssl_auth"] is True
+            assert description["services"][0]["url"] == 'http://petstore.swagger.io/v1'
 
     def test_run_configuration(self):
         urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
@@ -191,10 +202,10 @@ class TestXRDSST(IntegrationTestBase, IntegrationOpBase):
         self.step_enable_service_description(client_id)
         self.query_status()
 
-        self.step_add_service_access_rights()
+        self.step_add_service_access(client_id)
         self.query_status()
 
-        self.step_update_service_parameters()
+        self.step_update_service_parameters(client_id)
         configured_servers_at_end = self.query_status()
 
         assert_server_statuses_transitioned(unconfigured_servers_at_start, configured_servers_at_end)