+++ xrdsst/controllers/service.py
@@ -1,6 +1,3 @@
-from datetime import datetime
-
-import pytz
 from cement import ex
 from xrdsst.api import ClientsApi, ServiceDescriptionsApi, ServicesApi
 from xrdsst.api_client.api_client import ApiClient
@@ -18,7 +15,7 @@ class ServiceController(BaseController):
         stacked_type = 'nested'
         description = texts['service.controller.description']
 
-    @ex(label='add-description', help="Add service description", arguments=[])
+    @ex(help="Add service description", arguments=[])
     def add_description(self):
         active_config = self.load_config()
         full_op_path = self.op_path()
@@ -32,7 +29,7 @@ class ServiceController(BaseController):
 
         self.add_service_description(active_config)
 
-    @ex(label='enable-description', help="Enable service description", arguments=[])
+    @ex(help="Enable service description", arguments=[])
     def enable_description(self):
         active_config = self.load_config()
         full_op_path = self.op_path()
@@ -46,9 +43,19 @@ class ServiceController(BaseController):
 
         self.enable_service_description(active_config)
 
-    @ex(label='access-rights', help="Add access rights for service", arguments=[])
-    def add_rights(self):
-        self.add_access_rights(self.load_config())
+    @ex(help="Add access rights for service", arguments=[])
+    def add_access(self):
+        active_config = self.load_config()
+        full_op_path = self.op_path()
+
+        active_config, invalid_conf_servers = self.validate_op_config(active_config)
+        self.log_skipped_op_conf_invalid(invalid_conf_servers)
+
+        if not self.is_autoconfig():
+            active_config, unconfigured_servers = self.regroup_server_ops(active_config, full_op_path)
+            self.log_skipped_op_deps_unmet(full_op_path, unconfigured_servers)
+
+        self.add_access_rights(active_config)
 
     @ex(label='update-parameters', help="Update service parameters", arguments=[])
     def update_parameters(self):
@@ -103,7 +110,7 @@ class ServiceController(BaseController):
         code = service_description_conf['rest_service_code'] if service_description_conf['rest_service_code'] else None
         description_add = ServiceDescriptionAdd(url=service_description_conf['url'],
                                                 rest_service_code=code,
-                                                ignore_warnings=service_description_conf['ignore_warnings'],
+                                                ignore_warnings=True,
                                                 type=service_description_conf['type'])
         clients_api = ClientsApi(ApiClient(ss_configuration))
         try:
@@ -161,19 +168,27 @@ class ServiceController(BaseController):
                     if service_description:
                         for service in service_description.services:
                             try:
+                                client_id = None
                                 services_api = ServicesApi(ApiClient(ss_configuration))
-                                service_client = ServiceClient(id=client.id,
-                                                               name=client.member_name,
-                                                               service_client_type=ServiceClientType.SUBSYSTEM,
-                                                               rights_given_at=datetime.now(pytz.utc).isoformat())
-                                service_clients = ServiceClients(items=[service_client])
-                                response = services_api.add_service_service_clients(service.id, body=service_clients)
-                                if response:
-                                    BaseController.log_info("Added access rights for client '" + client.id +
-                                                            "' to use service '" + service.id + "' (got full id " + response[0].id + ")")
+                                access_list = service_description_conf["access"]
+                                for configurable_service in service_description_conf["services"]:
+                                    if service.service_code == configurable_service["service_code"]:
+                                        access_list = configurable_service["access"]
+                                for access in access_list:
+                                    client_id = client.instance_id + ':' + \
+                                                client.member_class + ':' + \
+                                                client.member_code + ':' + \
+                                                access
+                                    service_client = ServiceClient(id=client_id,
+                                                                   name=client.member_name,
+                                                                   service_client_type=ServiceClientType.SUBSYSTEM)
+                                    response = services_api.add_service_service_clients(service.id, body=ServiceClients(items=[service_client]))
+                                    if response:
+                                        BaseController.log_info("Added access rights for client '" + client_id +
+                                                                "' to use service '" + service.id + "' (full id " + response[0].id + ")")
                             except ApiException as err:
                                 if err.status == 409:
-                                    BaseController.log_info("Added access rights for client '" + client.id +
+                                    BaseController.log_info("Access rights for client '" + client_id +
                                                             "' to use service '" + service.id + "' already added")
                                 else:
                                     BaseController.log_api_error('ServicesApi->add_service_service_clients', err)
@@ -194,13 +209,26 @@ class ServiceController(BaseController):
                         for service in service_description.services:
                             try:
                                 services_api = ServicesApi(ApiClient(ss_configuration))
-                                service_update = ServiceUpdate(url=service_description_conf['url'],
-                                                               timeout=service_description_conf['timeout'],
-                                                               ssl_auth=service_description_conf['ssl_auth'],
-                                                               url_all=service_description_conf['url_all'],
-                                                               timeout_all=service_description_conf['timeout_all'],
-                                                               ssl_auth_all=service_description_conf['ssl_auth_all'],
-                                                               ignore_warnings=service_description_conf['ignore_warnings'])
+                                timeout = None
+                                timeout_all = service_description_conf["timeout_all"]
+                                ssl_auth = None
+                                ssl_auth_all = service_description_conf["ssl_auth_all"]
+                                url = None
+                                url_all = service_description_conf["url_all"]
+                                for configurable_service in service_description_conf["services"]:
+                                    if service.service_code == configurable_service["service_code"]:
+                                        timeout = configurable_service["timeout"]
+                                        timeout_all = None
+                                        ssl_auth = configurable_service["ssl_auth"]
+                                        ssl_auth_all = None
+                                        url = configurable_service["url"]
+                                        url_all = None
+                                service_update = ServiceUpdate(url=url,
+                                                               timeout=timeout,
+                                                               ssl_auth=ssl_auth,
+                                                               url_all=url_all,
+                                                               timeout_all=timeout_all,
+                                                               ssl_auth_all=ssl_auth_all)
                                 response = services_api.update_service(service.id, body=service_update)
                                 if response:
                                     BaseController.log_info("Updated service parameters for service '" + service.id +