+++ xrdsst/core/conf_keys.py
@@ -113,6 +113,26 @@ class ConfKeysSecServerClientServiceDesc:
     CONF_KEY_SS_CLIENT_SERVICE_DESC_URL = 'url'
     CONF_KEY_SS_CLIENT_SERVICE_DESC_REST_SERVICE_CODE = 'rest_service_code'
     CONF_KEY_SS_CLIENT_SERVICE_DESC_TYPE = 'type'
+    CONF_KEY_SS_CLIENT_SERVICE_DESC_CLIENT_ACCESS = 'access'
+    CONF_KEY_SS_CLIENT_SERVICE_DESC_URL_ALL = 'url_all'
+    CONF_KEY_SS_CLIENT_SERVICE_DESC_TIMEOUT_ALL = 'timeout_all'
+    CONF_KEY_SS_CLIENT_SERVICE_DESC_SSL_AUTH_ALL = 'ssl_auth_all'
+    CONF_KEY_SS_CLIENT_SERVICE_DESC_SERVICES = 'services'
+
+    @staticmethod
+    def descendant_conf_keys():
+        return [
+            (ConfKeysSecServerClients.CONF_KEY_SS_CLIENT_SERVICE_DESCS, ConfKeysSecServerClientServiceDescService)
+        ]
+
+
+# Known keys for xrdsst configuration file security server client service descriptions services configuration section.
+class ConfKeysSecServerClientServiceDescService:
+    CONF_KEY_SS_CLIENT_SERVICE_DESC_SERVICE_SERVICE_CODE = 'service_code'
+    CONF_KEY_SS_CLIENT_SERVICE_DESC_SERVICE_CLIENT_ACCESS = 'access'
+    CONF_KEY_SS_CLIENT_SERVICE_DESC_SERVICE_TIMEOUT = 'timeout'
+    CONF_KEY_SS_CLIENT_SERVICE_DESC_SERVICE_SSL_AUTH = 'ssl_auth'
+    CONF_KEY_SS_CLIENT_SERVICE_DESC_SERVICE_URL = 'url'
 
     @staticmethod
     def descendant_conf_keys():